Mov deploy and token to experimental.cli module
