+++ examples/deploy_run_echo.py
@@ -1,6 +1,7 @@
 import pathlib
 
 import pycape
+from pycape.experimental import cli
 
 CAPE_HOST = "wss://enclave.capeprivacy.com"
 ECHO_DEPLOY_PATH = pathlib.Path(__file__).parent.absolute() / "echo"
@@ -8,7 +9,7 @@ ECHO_DEPLOY_PATH = pathlib.Path(__file__).parent.absolute() / "echo"
 
 cape = pycape.Cape(url=CAPE_HOST)
 # Deploy Cape function
-function_ref = cape.deploy(ECHO_DEPLOY_PATH)
+function_ref = cli.deploy(ECHO_DEPLOY_PATH)
 message = cape.encrypt("Welcome to Cape".encode())
 # Run Cape function
 result = cape.run(function_ref, message)
